разбил на несколько классов
